Move get_profile_backend to views/users.py.
Fixes: #2710.
