+++ zerver/views/pointer.py
@@ -1,6 +1,5 @@
 from __future__ import absolute_import
 
-from django.core.exceptions import ValidationError
 from django.http import HttpRequest, HttpResponse
 from django.utils.translation import ugettext as _
 from six import text_type
@@ -9,8 +8,7 @@ from zerver.decorator import to_non_negative_int
 from zerver.lib.actions import do_update_pointer
 from zerver.lib.request import has_request_variables, JsonableError, REQ
 from zerver.lib.response import json_success
-from zerver.lib.utils import statsd, generate_random_token
-from zerver.models import UserProfile, Message, UserMessage
+from zerver.models import UserProfile, UserMessage
 
 def get_pointer_backend(request, user_profile):
     # type: (HttpRequest, UserProfile) -> HttpResponse
@@ -36,25 +34,3 @@ def update_pointer_backend(request, user_profile,
     do_update_pointer(user_profile, pointer, update_flags=update_flags)
 
     return json_success()
-
-def generate_client_id():
-    # type: () -> text_type
-    return generate_random_token(32)
-
-def get_profile_backend(request, user_profile):
-    # type: (HttpRequest, UserProfile) -> HttpResponse
-    result = dict(pointer        = user_profile.pointer,
-                  client_id      = generate_client_id(),
-                  max_message_id = -1,
-                  user_id        = user_profile.id,
-                  full_name      = user_profile.full_name,
-                  email          = user_profile.email,
-                  is_bot         = user_profile.is_bot,
-                  is_admin       = user_profile.is_realm_admin,
-                  short_name     = user_profile.short_name)
-
-    messages = Message.objects.filter(usermessage__user_profile=user_profile).order_by('-id')[:1]
-    if messages:
-        result['max_message_id'] = messages[0].id
-
-    return json_success(result)
